testing flyway vs liquibase

+++ sapp-db/src/main/resources/db/migration/v1/V1_0_0__Create_sequence_table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE sequence_table 
   ( 
      seq_name  VARCHAR(255), 
-     seq_value INTEGER 
+     seq_value BIGINT
   );
